requirements.txt
accelerate==0.25.0
huggingface-hub==0.20.1
Pillow==10.1.0
torch==2.3.1
torchvision==0.18.1
transformers==4.36.2
einops==0.7.0
gradio==4.15.0